drives to a pixel mark dependent on what mark the pixel is on.
Ability for pipeline to see the cube. Decisions can be made to drive to each spike mark. Untested.
