+++ TeamCode/build.gradle
@@ -32,6 +32,7 @@ android {
 
 dependencies {
     implementation project(':FtcRobotController')
+    implementation 'org.jetbrains:annotations:15.0'
     annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
     implementation 'org.openftc:easyopencv:1.7.0'
 }
